Refactor: turn twilio into its own script to be used in Flow
