+++ f/connectors/alerts/alerts_gcs.script.lock
@@ -24,5 +24,4 @@ requests==2.32.3
 rsa==4.9
 six==1.17.0
 tzdata==2024.2
-twilio==9.4.1
 urllib3==2.2.3
