+++ f/connectors/alerts/alerts_twilio.script.lock
@@ -1,0 +1,1 @@
+twilio==9.4.1
